refactor: rename context-mocks to state-mocks
