+++ packages/webview/src/component/dashboard/DashboardResourceCard.spec.ts
@@ -25,7 +25,7 @@ import { beforeEach, describe, expect, test, vi } from 'vitest';
 import KubernetesDashboardResourceCard from './DashboardResourceCard.svelte';
 import { FakeStateObject } from '/@/state/util/fake-state-object.svelte';
 import type { CurrentContextInfo } from '/@common/model/current-context-info';
-import { StatesMocks } from '/@/tests/context-mocks';
+import { StatesMocks } from '/@/tests/state-mocks';
 import { DependencyMocks } from '/@/tests/dependency-mocks';
 import type { ActiveResourcesCountInfo } from '/@common/model/active-resources-count-info';
 import type { ResourcesCountInfo } from '/@common/model/resources-count-info';
